Small DI refactoring.
